Finds the optimal Davies distribution for a dataset

Finds the best-fit Davies distribution using either the least-squares criterion (least.squares()) or maximum likelihood (maximum.likelihood())

least.squares(data, do.print = FALSE, start.v = NULL)
maximum.likelihood(data, do.print = FALSE, start.v = NULL)

Arguments

data

dataset to be fitted

do.print

Boolean with TRUE meaning print a GFM

start.v

A suitable starting vector of parameters c(C,lambda1,lambda2), with default NULL meaning to use start()

Details

Uses optim() to find the best-fit Davies distribution to a set of data.

Function least.squares() does not match that of Hankin and Lee 2006.

Value

Returns the parameters \(C,\lambda_1,\lambda_2\) of the best-fit Davies distribution to the dataset data

Author

Robin K. S. Hankin

See also

Note

BUGS:

Function least.squares() does not use the same methodology of Hankin and Lee 2006, and its use is discouraged pending implentation.

Quite apart from that, it can be screwed with bad value for start.v. Function maximum.likelihod() can be very slow. It might be possible to improve this by using some sort of hot-start for optim().

Examples

  p <- c(10 , 0.1 , 0.1)
  d <- rdavies(10,p)

  maximum.likelihood(d)  # quite slow
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> Warning: NaNs produced
#> [1] 8.90026962 0.04951972 0.27688159
  least.squares(d)       # much faster but not recommended
#> [1] 1.657780e+01 4.323839e-01 1.418071e-09